+++ lib/textus/cli/hook_group.rb
@@ -1,0 +1,9 @@
+module Textus
+  class CLI
+    class HookGroup < Group
+      self.cli_name = "hook"
+      subcommands["list"] = Hooks
+      subcommands["run"]  = Action
+    end
+  end
+end